backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.7.0
django-cuser==2017.3.16
django-filter==25.1
django-storages==1.14.6
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
django-modeltranslation==0.19.12
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
